+++ packages/minecraft/src/main/java/minechain/exchange/Exchange.java
@@ -1,11 +1,11 @@
-package minechain.channels;
+package minechain.exchange;
 
 import com.rabbitmq.client.DeliverCallback;
 import com.rabbitmq.client.Delivery;
 import java.lang.reflect.InvocationTargetException;
 import minechain.Rabbit;
 
-public class Exchange {
+public abstract class Exchange {
 
   private String exchange;
   private String type;
@@ -21,7 +21,7 @@ public class Exchange {
     this.channel = Rabbit.getInstance();
 
     for (var m : this.getClass().getMethods()) {
-      var routeMethod = (Route) m.getAnnotation(Route.class);
+      var routeMethod = m.getAnnotation(Route.class);
       var self = this;
       if (routeMethod != null) {
         DeliverCallback deliverCallback = new DeliverCallback() {
