+++ packages/minecraft/src/main/java/minechain/exchange/RegistrationExchange.java
@@ -1,4 +1,4 @@
-package minechain.channels;
+package minechain.exchange;
 
 import com.google.gson.Gson;
 import com.rabbitmq.client.Delivery;
