+++ packages/minecraft/src/main/java/minechain/exchange/Route.java
@@ -1,4 +1,4 @@
-package minechain.channels;
+package minechain.exchange;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
